requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
python-decouple==3.8
yandex-music==2.2.0